% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.guardduty_operations.R
\name{archive_findings}
\alias{archive_findings}
\title{Archives Amazon GuardDuty findings specified by the list of finding IDs}
\usage{
archive_findings(DetectorId, FindingIds)
}
\arguments{
\item{DetectorId}{The ID of the detector that specifies the GuardDuty service whose findings you want to archive.}

\item{FindingIds}{IDs of the findings that you want to archive.}
}
\description{
Archives Amazon GuardDuty findings specified by the list of finding IDs.
}
\examples{

}
