% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.guardduty_operations.R
\name{create_filter}
\alias{create_filter}
\title{Creates a filter using the specified finding criteria}
\usage{
create_filter(DetectorId, FindingCriteria, Name, Action = NULL,
  ClientToken = NULL, Description = NULL, Rank = NULL)
}
\arguments{
\item{DetectorId}{The unique ID of the detector that you want to update.}

\item{FindingCriteria}{Represents the criteria to be used in the filter for querying findings.}

\item{Name}{The name of the filter.}

\item{Action}{Specifies the action that is to be applied to the findings that match the filter.}

\item{ClientToken}{The idempotency token for the create request.}

\item{Description}{The description of the filter.}

\item{Rank}{Specifies the position of the filter in the list of current filters. Also specifies the order in which this filter is applied to the findings.}
}
\description{
Creates a filter using the specified finding criteria.
}
\examples{

}
