% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.guardduty_operations.R
\name{create_threat_intel_set}
\alias{create_threat_intel_set}
\title{Create a new ThreatIntelSet}
\usage{
create_threat_intel_set(DetectorId, Format, Activate, Location, Name,
  ClientToken = NULL)
}
\arguments{
\item{DetectorId}{The unique ID of the detector that you want to update.}

\item{Format}{The format of the file that contains the ThreatIntelSet.}

\item{Activate}{A boolean value that indicates whether GuardDuty is to start using the uploaded ThreatIntelSet.}

\item{Location}{The URI of the file that contains the ThreatIntelSet. For example (https://s3.us-west-2.amazonaws.com/my-bucket/my-object-key).}

\item{Name}{A user-friendly ThreatIntelSet name that is displayed in all finding generated by activity that involves IP addresses included in this ThreatIntelSet.}

\item{ClientToken}{The idempotency token for the create request.}
}
\description{
Create a new ThreatIntelSet. ThreatIntelSets consist of known malicious IP addresses. GuardDuty generates findings based on ThreatIntelSets.
}
\examples{

}
