% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.guardduty_operations.R
\name{decline_invitations}
\alias{decline_invitations}
\title{Declines invitations sent to the current member account by AWS account specified by their account IDs}
\usage{
decline_invitations(AccountIds)
}
\arguments{
\item{AccountIds}{A list of account IDs of the AWS accounts that sent invitations to the current member account that you want to decline invitations from.}
}
\description{
Declines invitations sent to the current member account by AWS account specified by their account IDs.
}
\examples{

}
