% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.guardduty_operations.R
\name{delete_detector}
\alias{delete_detector}
\title{Deletes a Amazon GuardDuty detector specified by the detector ID}
\usage{
delete_detector(DetectorId)
}
\arguments{
\item{DetectorId}{[required] The unique ID that specifies the detector that you want to delete.}
}
\description{
Deletes a Amazon GuardDuty detector specified by the detector ID.
}
\section{Accepted Parameters}{
\preformatted{delete_detector(
  DetectorId = "string"
)
}
}

