% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.guardduty_operations.R
\name{delete_filter}
\alias{delete_filter}
\title{Deletes the filter specified by the filter name}
\usage{
delete_filter(DetectorId, FilterName)
}
\arguments{
\item{DetectorId}{The unique ID that specifies the detector where you want to delete a filter.}

\item{FilterName}{The name of the filter.}
}
\description{
Deletes the filter specified by the filter name.
}
\examples{

}
