% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.guardduty_operations.R
\name{delete_ip_set}
\alias{delete_ip_set}
\title{Deletes the IPSet specified by the IPSet ID}
\usage{
delete_ip_set(
  DetectorId = "string",
  IpSetId = "string"
)
}
\arguments{
\item{DetectorId}{[required] The detectorID that specifies the GuardDuty service whose IPSet you want to delete.}

\item{IpSetId}{[required] The unique ID that specifies the IPSet that you want to delete.}
}
\description{
Deletes the IPSet specified by the IPSet ID.
}
\examples{

}
