% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.guardduty_operations.R
\name{disassociate_from_master_account}
\alias{disassociate_from_master_account}
\title{Disassociates the current GuardDuty member account from its master account}
\usage{
disassociate_from_master_account(DetectorId)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the detector of the GuardDuty member account.}
}
\description{
Disassociates the current GuardDuty member account from its master account.
}
\section{Accepted Parameters}{
\preformatted{disassociate_from_master_account(
  DetectorId = "string"
)
}
}

