% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.guardduty_operations.R
\name{disassociate_members}
\alias{disassociate_members}
\title{Disassociates GuardDuty member accounts (to the current GuardDuty master account) specified by the account IDs}
\usage{
disassociate_members(DetectorId, AccountIds)
}
\arguments{
\item{DetectorId}{The unique ID of the detector of the GuardDuty account whose members you want to disassociate from master.}

\item{AccountIds}{A list of account IDs of the GuardDuty member accounts that you want to disassociate from master.}
}
\description{
Disassociates GuardDuty member accounts (to the current GuardDuty master account) specified by the account IDs.
}
\examples{

}
