% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.guardduty_operations.R
\name{get_detector}
\alias{get_detector}
\title{Retrieves an Amazon GuardDuty detector specified by the detectorId}
\usage{
get_detector(DetectorId)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the detector that you want to retrieve.}
}
\description{
Retrieves an Amazon GuardDuty detector specified by the detectorId.
}
\section{Accepted Parameters}{
\preformatted{get_detector(
  DetectorId = "string"
)
}
}

