% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.guardduty_operations.R
\name{get_filter}
\alias{get_filter}
\title{Returns the details of the filter specified by the filter name}
\usage{
get_filter(DetectorId, FilterName)
}
\arguments{
\item{DetectorId}{[required] The detector ID that specifies the GuardDuty service where you want to list the details of the specified filter.}

\item{FilterName}{[required] The name of the filter whose details you want to get.}
}
\description{
Returns the details of the filter specified by the filter name.
}
\section{Accepted Parameters}{
\preformatted{get_filter(
  DetectorId = "string",
  FilterName = "string"
)
}
}

