% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.guardduty_operations.R
\name{get_findings_statistics}
\alias{get_findings_statistics}
\title{Lists Amazon GuardDuty findings' statistics for the specified detector ID}
\usage{
get_findings_statistics(DetectorId, FindingStatisticTypes,
  FindingCriteria = NULL)
}
\arguments{
\item{DetectorId}{The ID of the detector that specifies the GuardDuty service whose findings' statistics you want to retrieve.}

\item{FindingStatisticTypes}{Types of finding statistics to retrieve.}

\item{FindingCriteria}{Represents the criteria used for querying findings.}
}
\description{
Lists Amazon GuardDuty findings' statistics for the specified detector ID.
}
\examples{

}
