% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.guardduty_operations.R
\name{get_invitations_count}
\alias{get_invitations_count}
\title{Returns the count of all GuardDuty membership invitations that were sent to the current member account except the currently accepted invitation}
\usage{
get_invitations_count()
}
\description{
Returns the count of all GuardDuty membership invitations that were sent to the current member account except the currently accepted invitation.
}
\examples{

}
