% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.guardduty_operations.R
\name{get_ip_set}
\alias{get_ip_set}
\title{Retrieves the IPSet specified by the IPSet ID}
\usage{
get_ip_set(
  DetectorId = "string",
  IpSetId = "string"
)
}
\arguments{
\item{DetectorId}{[required] The detectorID that specifies the GuardDuty service whose IPSet you want to retrieve.}

\item{IpSetId}{[required] The unique ID that specifies the IPSet that you want to describe.}
}
\description{
Retrieves the IPSet specified by the IPSet ID.
}
\examples{

}
