% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.guardduty_operations.R
\name{get_master_account}
\alias{get_master_account}
\title{Provides the details for the GuardDuty master account to the current GuardDuty member account}
\usage{
get_master_account(DetectorId)
}
\arguments{
\item{DetectorId}{The unique ID of the detector of the GuardDuty member account.}
}
\description{
Provides the details for the GuardDuty master account to the current GuardDuty member account.
}
\examples{

}
