% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.guardduty_operations.R
\name{get_threat_intel_set}
\alias{get_threat_intel_set}
\title{Retrieves the ThreatIntelSet that is specified by the ThreatIntelSet ID}
\usage{
get_threat_intel_set(ThreatIntelSetId, DetectorId)
}
\arguments{
\item{ThreatIntelSetId}{The unique ID that specifies the ThreatIntelSet that you want to describe.}

\item{DetectorId}{The detectorID that specifies the GuardDuty service whose ThreatIntelSet you want to describe.}
}
\description{
Retrieves the ThreatIntelSet that is specified by the ThreatIntelSet ID.
}
\examples{

}
