% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.guardduty_operations.R
\name{list_filters}
\alias{list_filters}
\title{Returns a paginated list of the current filters}
\usage{
list_filters(
  DetectorId = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
\arguments{
\item{DetectorId}{[required] The ID of the detector that specifies the GuardDuty service where you want to list filters.}

\item{MaxResults}{Indicates the maximum number of items that you want in the response. The maximum value is 50.}

\item{NextToken}{Paginates results. Set the value of this parameter to NULL on your first call to the ListFilters operation.For subsequent calls to the operation, fill nextToken in the request with the value of nextToken from the previous response to continue listing data.}
}
\description{
Returns a paginated list of the current filters.
}
\examples{

}
