% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.guardduty_package.R
\docType{package}
\name{paws.guardduty-package}
\alias{paws.guardduty}
\alias{paws.guardduty-package}
\title{paws.guardduty: Amazon GuardDuty}
\description{
Assess, monitor, manage, and remediate security
    issues across your AWS infrastructure, applications, and data.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
