% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.guardduty_operations.R
\name{unarchive_findings}
\alias{unarchive_findings}
\title{Unarchives Amazon GuardDuty findings specified by the list of finding IDs}
\usage{
unarchive_findings(DetectorId, FindingIds)
}
\arguments{
\item{DetectorId}{The ID of the detector that specifies the GuardDuty service whose findings you want to unarchive.}

\item{FindingIds}{IDs of the findings that you want to unarchive.}
}
\description{
Unarchives Amazon GuardDuty findings specified by the list of finding IDs.
}
\examples{

}
