% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.guardduty_operations.R
\name{update_detector}
\alias{update_detector}
\title{Updates an Amazon GuardDuty detector specified by the detectorId}
\usage{
update_detector(DetectorId, Enable = NULL,
  FindingPublishingFrequency = NULL)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the detector that you want to update.}

\item{Enable}{Updated boolean value for the detector that specifies whether the detector is enabled.}

\item{FindingPublishingFrequency}{A enum value that specifies how frequently customer got Finding updates published.}
}
\description{
Updates an Amazon GuardDuty detector specified by the detectorId.
}
\section{Accepted Parameters}{
\preformatted{update_detector(
  DetectorId = "string",
  Enable = TRUE|FALSE,
  FindingPublishingFrequency = "FIFTEEN_MINUTES"|"ONE_HOUR"|"SIX_HOURS"
)
}
}

