% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.guardduty_operations.R
\name{update_filter}
\alias{update_filter}
\title{Updates the filter specified by the filter name}
\usage{
update_filter(DetectorId, FilterName, Action = NULL, Description = NULL,
  FindingCriteria = NULL, Rank = NULL)
}
\arguments{
\item{DetectorId}{The unique ID of the detector that specifies the GuardDuty service where you want to update a filter.}

\item{FilterName}{The name of the filter.}

\item{Action}{Specifies the action that is to be applied to the findings that match the filter.}

\item{Description}{The description of the filter.}

\item{FindingCriteria}{Represents the criteria to be used in the filter for querying findings.}

\item{Rank}{Specifies the position of the filter in the list of current filters. Also specifies the order in which this filter is applied to the findings.}
}
\description{
Updates the filter specified by the filter name.
}
\examples{

}
