% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.guardduty_operations.R
\name{update_findings_feedback}
\alias{update_findings_feedback}
\title{Marks specified Amazon GuardDuty findings as useful or not useful}
\usage{
update_findings_feedback(DetectorId, Feedback, FindingIds, Comments = NULL)
}
\arguments{
\item{DetectorId}{The ID of the detector that specifies the GuardDuty service whose findings you want to mark as useful or not useful.}

\item{Feedback}{Valid values: USEFUL | NOT_USEFUL}

\item{FindingIds}{IDs of the findings that you want to mark as useful or not useful.}

\item{Comments}{Additional feedback about the GuardDuty findings.}
}
\description{
Marks specified Amazon GuardDuty findings as useful or not useful.
}
\examples{

}
