% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.guardduty_operations.R
\name{update_ip_set}
\alias{update_ip_set}
\title{Updates the IPSet specified by the IPSet ID}
\usage{
update_ip_set(Activate = NULL, DetectorId, IpSetId, Location = NULL,
  Name = NULL)
}
\arguments{
\item{Activate}{The updated boolean value that specifies whether the IPSet is active or not.}

\item{DetectorId}{[required] The detectorID that specifies the GuardDuty service whose IPSet you want to update.}

\item{IpSetId}{[required] The unique ID that specifies the IPSet that you want to update.}

\item{Location}{The updated URI of the file that contains the IPSet. For example (https://s3.us-west-2.amazonaws.com/my-bucket/my-object-key).}

\item{Name}{The unique ID that specifies the IPSet that you want to update.}
}
\description{
Updates the IPSet specified by the IPSet ID.
}
\section{Accepted Parameters}{
\preformatted{update_ip_set(
  Activate = TRUE|FALSE,
  DetectorId = "string",
  IpSetId = "string",
  Location = "string",
  Name = "string"
)
}
}

