% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.guardduty_operations.R
\name{accept_invitation}
\alias{accept_invitation}
\title{Accepts the invitation to be monitored by a master GuardDuty account}
\usage{
accept_invitation(DetectorId, MasterId, InvitationId)
}
\arguments{
\item{DetectorId}{The unique ID of the detector of the GuardDuty member account.}

\item{MasterId}{The account ID of the master GuardDuty account whose invitation you're accepting.}

\item{InvitationId}{This value is used to validate the master account to the member account.}
}
\description{
Accepts the invitation to be monitored by a master GuardDuty account.
}
\examples{

}
