% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.guardduty_operations.R
\name{delete_filter}
\alias{delete_filter}
\title{Deletes the filter specified by the filter name}
\usage{
delete_filter(DetectorId, FilterName)
}
\arguments{
\item{DetectorId}{[required] The unique ID that specifies the detector where you want to delete a filter.}

\item{FilterName}{[required] The name of the filter.}
}
\description{
Deletes the filter specified by the filter name.
}
\section{Accepted Parameters}{
\preformatted{delete_filter(
  DetectorId = "string",
  FilterName = "string"
)
}
}

