% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.guardduty_operations.R
\name{delete_threat_intel_set}
\alias{delete_threat_intel_set}
\title{Deletes ThreatIntelSet specified by the ThreatIntelSet ID}
\usage{
delete_threat_intel_set(DetectorId, ThreatIntelSetId)
}
\arguments{
\item{DetectorId}{[required] The detectorID that specifies the GuardDuty service whose ThreatIntelSet you want to delete.}

\item{ThreatIntelSetId}{[required] The unique ID that specifies the ThreatIntelSet that you want to delete.}
}
\description{
Deletes ThreatIntelSet specified by the ThreatIntelSet ID.
}
\section{Accepted Parameters}{
\preformatted{delete_threat_intel_set(
  DetectorId = "string",
  ThreatIntelSetId = "string"
)
}
}

